requirements.txt
Pillow>=12.0.0
opencv-python>=4.12.0
numpy>=2.2.0
